added first attempt at background plotting
first cut does not follow the hardware, it just follows the logical behaviour. So although inefficient it does seem to work and even allows donkey kong to run for a bit! :)
(this is because it seems complicated, will have to read docs properly)

next step: understand hardware and try to follow it

commented out APU panic for the moment