Fix bug caused by previous submit
Trying to evaluate sprites for line0 on the last line was
a mistake as it seems line 0 does not have sprites. Instead,
just clear the buffers before going to the pre-render line.

Source:
https://wiki.nesdev.com/w/index.php/PPU_sprite_evaluation:
Sprite evaluation does not happen on the pre-render scanline.
Because evaluation applies to the next line's sprite rendering,
no sprites will be rendered on the first scanline, and this is
why there is a 1 line offset on a sprite's Y coordinate.