save battery backed ram to file
Saved on the home directory within .config/gones.
Using the name of the iNES file and the hash of prgRom in the save
filename - this is because I tend to use the same iNES.
name, eg /tmp/a.nes so I don't have to keep changing the debug config
Games are saved automatically on exit.
todo: save on a timer?