Save and Load State
Now we can save/load state:
save -> we save the state of the whole emulator (or mostly) into a save file.
This means we recursively go through all data structures in the nes serialising
them into a file via a Serialiser.
load -> well, it's loading back of the saved state...

To achieve this we introduce new interfaces Serialiser and Serialisable. Every
state information in the nes that needs to persist when saving/loading has been
made Serialisable and Serialiser is the interface which glues it to a specific
Serialiser. Eg atm we're using the gob serialiser

All data members of strucutures must be manually inserted into a serialise call.
I've tried to make this all automated but I had trouble doing the deserialise part
of it since reflection does not work well on private fields :(
Not all is bad news, because using reflection we're able to automatically call
Serialise on the non native types, allowing us to have a single call per structure!

Warning: No plans to introduce versioning support yet... and so the saves on today's
emulator might not be loaded on the next update :(